[feat] 스탬프 관련 API 변경 사항 반영
