+++ core/data/src/main/kotlin/com/unifest/android/core/data/mapper/MyWaitingMapper.kt
@@ -1,7 +1,7 @@
 package com.unifest.android.core.data.mapper
 
 import com.unifest.android.core.model.MyWaitingModel
-import com.unifest.android.core.network.response.MyWaiting
+import com.unifest.android.core.network.response.waiting.MyWaiting
 
 internal fun MyWaiting.toModel(): MyWaitingModel {
     return MyWaitingModel(
