+++ core/data/src/main/kotlin/com/unifest/android/core/data/repository/StampRepository.kt
@@ -1,9 +1,10 @@
 package com.unifest.android.core.data.repository
 
 import com.unifest.android.core.model.StampBoothModel
+import com.unifest.android.core.model.StampRecordModel
 
 interface StampRepository {
-    suspend fun getCollectedStampCount(): Result<Int>
+    suspend fun getCollectedStampCount(): Result<List<StampRecordModel>>
     suspend fun getStampEnabledBoothList(): Result<List<StampBoothModel>>
     suspend fun registerStamp(boothId: Long): Result<Unit>
 }
