+++ core/model/src/main/kotlin/com/unifest/android/core/model/StampRecordModel.kt
@@ -1,0 +1,7 @@
+package com.unifest.android.core.model
+
+data class StampRecordModel(
+    val stampRecordId: Long,
+    val stampInfoId: Long,
+    val deviceId: String,
+)
