--- core/network/src/main/kotlin/com/unifest/android/core/network/response/CollectedStampCountResponse.kt
@@ -1,14 +1,0 @@
-package com.unifest.android.core.network.response
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class CollectedStampCountResponse(
-    @SerialName("code")
-    val code: String,
-    @SerialName("message")
-    val message: String,
-    @SerialName("data")
-    val data: Int,
-)
